Add wafo\.libs to the DLL search path
The compiled Fortran extensions are placed in wafo\.libs as dll files while corresponding pyd files are in the wafo root. For the pyd files to be able to locate the dlls on import we need to add their path with os.add_dll_directory(path) which is the supported way since Python 3.8. However, for me that method only works in Python 3.10. To make it work for Python<3.10, wafo\.libs is also added to the PATH environment variable. Tested on Windows 10 with Python 3.6 up to 3.10.